Don't protect 'rm' when using vscode
The protection interrupts agentic AI's ability to work

+++ etc/profile.d/0-core.sh
@@ -6,7 +6,9 @@
 #
 
 # aliases
-alias rm='rm -i'
+if [ "${TERM_PROGRAM:-}" != "vscode" ]; then
+    alias rm='rm -i'
+fi
 alias view="vim -R"
 
 # environment
